feat: add Exceptions and Error page

+++ src/main/java/br/com/fiap/challenge/Challenge01/services/ClinicaService.java
@@ -3,6 +3,8 @@ package br.com.fiap.challenge.Challenge01.services;
 import br.com.fiap.challenge.Challenge01.dto.clinica.DtoAtualizarClinica;
 import br.com.fiap.challenge.Challenge01.dto.clinica.DtoCriarClinica;
 import br.com.fiap.challenge.Challenge01.dto.clinica.DtoListarClinica;
+import br.com.fiap.challenge.Challenge01.exceptions.ConflictException;
+import br.com.fiap.challenge.Challenge01.exceptions.ObjectNotFoundException;
 import br.com.fiap.challenge.Challenge01.models.Clinica;
 import br.com.fiap.challenge.Challenge01.models.Endereco;
 import br.com.fiap.challenge.Challenge01.repositories.ClinicaRepository;
@@ -25,27 +27,23 @@ public class ClinicaService {
 
     @Transactional
     public DtoListarClinica createClinica(DtoCriarClinica dados) {
-        if (!clinicaRepository.existsByCnpj(dados.cnpj())){
-            var endereco = enderecoRepository.save(new Endereco(dados.endereco()));
-            var clinica = new Clinica(dados);
-            clinica.setEndereco(endereco);
-            this.save(clinica);
+        if (clinicaRepository.existsByCnpj(dados.cnpj())) throw new ConflictException("CNPJ already registered");
 
-            return new DtoListarClinica(clinica);
-        }
-        return null;
+        var endereco = enderecoRepository.save(new Endereco(dados.endereco()));
+        var clinica = new Clinica(dados);
+        clinica.setEndereco(endereco);
+        this.save(clinica);
+        return new DtoListarClinica(clinica);
     }
 
     @Transactional
     public DtoListarClinica updateClinica(Long id, @Valid DtoAtualizarClinica dados) {
-        if (clinicaRepository.existsById(id)){
-            var clinica = clinicaRepository.getReferenceById(id);
-            clinica.atualizarClinica(dados);
-            this.save(clinica);
+        if (!clinicaRepository.existsById(id)) throw new ObjectNotFoundException("Clinica not found");
+        var clinica = clinicaRepository.getReferenceById(id);
+        clinica.atualizarClinica(dados);
+        this.save(clinica);
 
-            return new DtoListarClinica(clinica);
-        }
-        return null;
+        return new DtoListarClinica(clinica);
     }
 
     @Transactional
@@ -56,25 +54,21 @@ public class ClinicaService {
     @Transactional
     public DtoListarClinica getClinicaByCnpj(String cnpj) {
         var clinica = clinicaRepository.findByCnpj(cnpj);
-        if (clinica != null) {
-
-            return new DtoListarClinica(clinica);
-        }
-        return null;
+        if (clinica == null) throw new ObjectNotFoundException("Clinica not found");
+        return new DtoListarClinica(clinica);
     }
 
     @Transactional
     public void deleteById(Long id) {
+        if (!clinicaRepository.existsById(id)) throw new ObjectNotFoundException("Clinica not found");
         clinicaRepository.deleteById(id);
     }
 
     @Transactional
     public DtoListarClinica findById(Long id) {
-        Optional<Clinica> clinica = clinicaRepository.findById(id);
-        if (clinica.isPresent()){
-            return new DtoListarClinica(clinica.get());
-        }
-        throw new RuntimeException("id not found");
+        return clinicaRepository.findById(id)
+                .map(DtoListarClinica::new)
+                .orElseThrow(() -> new ObjectNotFoundException("Clinica not found"));
     }
 
     @Transactional
