+++ src/main/java/br/com/fiap/challenge/Challenge01/services/PacienteService.java
@@ -3,6 +3,9 @@ package br.com.fiap.challenge.Challenge01.services;
 import br.com.fiap.challenge.Challenge01.dto.paciente.DtoAtualizarPaciente;
 import br.com.fiap.challenge.Challenge01.dto.paciente.DtoCriarPaciente;
 import br.com.fiap.challenge.Challenge01.dto.paciente.DtoListarPaciente;
+import br.com.fiap.challenge.Challenge01.exceptions.ConflictException;
+import br.com.fiap.challenge.Challenge01.exceptions.InvalidDataException;
+import br.com.fiap.challenge.Challenge01.exceptions.ObjectNotFoundException;
 import br.com.fiap.challenge.Challenge01.models.Endereco;
 import br.com.fiap.challenge.Challenge01.models.Paciente;
 import br.com.fiap.challenge.Challenge01.repositories.EnderecoRepository;
@@ -13,8 +16,8 @@ import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 import org.springframework.web.bind.annotation.RequestBody;
 
+import java.time.LocalDate;
 import java.util.List;
-import java.util.Optional;
 
 @Service
 public class PacienteService {
@@ -26,15 +29,14 @@ public class PacienteService {
 
     @Transactional
     public DtoListarPaciente createPaciente(@Valid @RequestBody DtoCriarPaciente dados) {
-        if (!pacienteRepository.existsByCpf(dados.cpf())){
-            var endereco = enderecoRepository.save(new Endereco(dados.endereco()));
-            var paciente = new Paciente(dados);
-            paciente.setEndereco(endereco);
-            this.save(paciente);
+        if (pacienteRepository.existsByCpf(dados.cpf())) throw new ConflictException("Cpf already registered");
+        if (dados.dataNascimento().isAfter(LocalDate.now())) throw new InvalidDataException();
 
-            return new DtoListarPaciente(paciente);
-        }
-        return null;
+        var endereco = enderecoRepository.save(new Endereco(dados.endereco()));
+        var paciente = new Paciente(dados);
+        paciente.setEndereco(endereco);
+        this.save(paciente);
+        return new DtoListarPaciente(paciente);
     }
 
     @Transactional
@@ -45,31 +47,20 @@ public class PacienteService {
     @Transactional
     public DtoListarPaciente getPacienteByCpf(String cpf) {
         var paciente = pacienteRepository.findByCpf(cpf);
-        if (paciente != null) {
-            return new DtoListarPaciente(paciente);
-        }
-        return null;
+
+        if (paciente == null) throw new ObjectNotFoundException("Pacient not found");
+
+        return new DtoListarPaciente(paciente);
     }
 
     @Transactional
     public DtoListarPaciente updatePaciente(String cpf, DtoAtualizarPaciente dados) {
-        if (pacienteRepository.existsByCpf(cpf)){
-            var paciente = pacienteRepository.getReferenceByCpf(cpf);
-            paciente.atualizarPaciente(dados);
-            this.save(paciente);
+        if (!pacienteRepository.existsByCpf(cpf)) throw new ObjectNotFoundException("Pacient not found");
 
-            return new DtoListarPaciente(paciente);
-        }
-        return null;
-    }
-
-    @Transactional
-    public DtoListarPaciente findByCpf(String cpf) {
-        Optional<Paciente> paciente = Optional.ofNullable(pacienteRepository.findByCpf(cpf));
-        if (paciente.isPresent()){
-            return new DtoListarPaciente(paciente.get());
-        }
-        throw new RuntimeException("id not found");
+        var paciente = pacienteRepository.getReferenceByCpf(cpf);
+        paciente.atualizarPaciente(dados);
+        this.save(paciente);
+        return new DtoListarPaciente(paciente);
     }
 
     @Transactional
@@ -79,6 +70,7 @@ public class PacienteService {
 
     @Transactional
     public void deleteById(Long id) {
+        if (!pacienteRepository.existsById(id)) throw new ObjectNotFoundException("Pacient not found");
         pacienteRepository.deleteById(id);
     }
 }
